holocron/backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
pgvector==0.3.6

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# FSRS
numpy==2.2.1
